ロック画面コントロールを長押ししてもアプリが開かない問題を修正
