+++ ios/RideSessionActivity/Schemes/Prod/Info.plist
@@ -8,6 +8,8 @@
 	<string>$(MARKETING_VERSION)</string>
 	<key>CFBundleVersion</key>
 	<string>$(CURRENT_PROJECT_VERSION)</string>
+	<key>CURRENT_SCHEME_NAME</key>
+	<string>ProdTrainLCD</string>
 	<key>NSExtension</key>
 	<dict>
 		<key>NSExtensionPointIdentifier</key>
